[Wasm] Move some functions to `WasmBinaryGenerationPipelinePhase`
…and make them private.

This is to discourage calling them directly from test code.
Test code should use pipeline phases, to make sure that we're testing
the same behavior as what happens in production.

KT-77029
